requirements.txt
streamlit
kubernetes
langchain
langchain-cohere
python-dotenv